+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 [Full Changelog](https://github.com/whipper-team/whipper/compare/v0.10.0...HEAD)
 
+- Handle FreeBSD drive status, mounts, tray control, and cdrdao output (#686).
+
 ## [v0.10.0](https://github.com/whipper-team/whipper/tree/v0.10.0) (2021-05-17)
 
 [Full Changelog](https://github.com/whipper-team/whipper/compare/v0.9.0...v0.10.0)